requirements.txt
pyautogui
Pillow 
